// src/api/generated/models/ImageMetadataDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ThumbnailBoundingBoxDto } from './ThumbnailBoundingBoxDto';

export type ImageMetadataDto = {
    modelName: string;
    boundingBox: ThumbnailBoundingBoxDto;
    colorLegend: Record<string, string>;
};